DRY isotest code
