+++ gulp/isotest.js
@@ -4,15 +4,11 @@
 "use strict";
 
 module.exports = (gulp, plugins, blueprint) => {
+    const path = require("path");
     const rs = require("run-sequence").use(gulp);
-    const mocha = require("gulp-mocha");
 
-    blueprint.projectsWithBlock("isotest").forEach((project) => {
-        gulp.task(`isotest-${project.id}`, [`typescript-compile-${project.id}`], (done) => {
-            return gulp.src(project.cwd + "test.iso/**/*")
-                .pipe(mocha());
-        });
+    blueprint.task("isotest", "mocha", ["typescript-compile-*"], (project) => {
+        return gulp.src(path.join(project.cwd, "test.iso", "index.js"))
+            .pipe(plugins.mocha());
     });
-
-    gulp.task("isotest", (done) => rs(...blueprint.taskMapper("isotest", "isotest-"), done));
 };
